flask-backend/requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9

pytest>=7.0.0
pytest-mock>=3.6.1
